Made the union of a similar functional
