+++ Source/AI/TexasHoldem.AI.SmartPlayer/SmartPlayer.cs
@@ -16,7 +16,7 @@
 
         public override PlayerAction PostingBlind(IPostingBlindContext context)
         {
-            throw new NotImplementedException();
+            return context.BlindAction;
         }
 
         public override PlayerAction GetTurn(IGetTurnContext context)
