+++ Source/TexasHoldem.Logic/TexasHoldem.Logic.csproj
@@ -49,19 +49,13 @@
     <Compile Include="Cards\IDeck.cs" />
     <Compile Include="Extensions\EnumerableExtensions.cs" />
     <Compile Include="Extensions\RandomProvider.cs" />
-    <Compile Include="GameMechanics\BaseBettingLogic.cs" />
-    <Compile Include="GameMechanics\BaseHandLogic.cs" />
-    <Compile Include="GameMechanics\BaseTexasHoldemGame.cs" />
+    <Compile Include="GameMechanics\BettingLogic.cs" />
+    <Compile Include="GameMechanics\HandLogic.cs" />
+    <Compile Include="GameMechanics\TexasHoldemGame.cs" />
     <Compile Include="GameMechanics\IHandLogic.cs" />
     <Compile Include="GameMechanics\IInternalPlayer.cs" />
     <Compile Include="GameMechanics\InternalPlayerMoney.cs" />
-    <Compile Include="GameMechanics\MultiplePlayersBettingLogic.cs" />
-    <Compile Include="GameMechanics\MultiplePlayersHandLogic.cs" />
-    <Compile Include="GameMechanics\MultiplePlayersTexasHoldemGame.cs" />
-    <Compile Include="GameMechanics\HeadsUpBettingLogic.cs" />
-    <Compile Include="GameMechanics\HeadsUpHandLogic.cs" />
     <Compile Include="GameMechanics\ITexasHoldemGame.cs" />
-    <Compile Include="GameMechanics\HeadsUpTexasHoldemGame.cs" />
     <Compile Include="GameMechanics\SidePot.cs" />
     <Compile Include="GameRoundType.cs" />
     <Compile Include="Helpers\BestHand.cs" />
